third_party/blink/web_tests/html5lib/generated/run-tests1-data-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
../resources/tests1.dat: PASS
